PhasicsDensityAnalysis: iterative wavefront reconstruction by FT
For the _reconstruct_wavefront_FT_from_gradient_FTs(), or 'velghe' method
of wavefront reconstruction from wavefront gradients, the FT of the
wavefront is estimated from the FTs of the gradients. If the border of
the wavefront (to be estimated) image isn't flat, such as in the case of
a plume in a neutral gas image, edge effects cause reflection in the FTs.

This commit adds padding and an iterative procedure as given by
   Roddier, François, and Claude Roddier. "Wavefront reconstruction using
   iterative Fourier transforms." Applied Optics 30, no. 11 (1991): 1325-1327
that estimates the continuation of the wavefront gradient in order to
mitigate edge effects.